ENGR00175360 Camera preview would hang after recording for many times.
The root cause was in the process of stopping recording, there was more
frame sent to VPU. In this condition, CameraService would drop this frame.
But camera HAL thought it already sent to VPU and would not release it anymore.
The way to fix it, after stopping record, release the frame sent to VPU in
camera HAL.

Signed-off-by: YangYonggang <b31664@freescale.com>